gcp-deployment/requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
werkzeug==2.3.7
python-dotenv==1.0.0
pydantic>=2.10,<3

# Google Cloud Functions specific
functions-framework==3.2.0

# API clients
httpx>=0.27,<1
aiohttp==3.8.5

# AI and NLP
pydantic-ai==0.0.32
openai>==1.3.7
langchain==0.0.312

# Data processing
numpy==1.24.4
pandas==2.0.3

# Caching and performance
cachetools==5.3.1
prometheus-client==0.17.1

# Utilities
tenacity==8.2.3
loguru==0.7.2
